requirements.txt
Flask<1
# applicationinsights
adal
requests
azure-mgmt-resource==0.31.0
haikunator
msrestazure